+++ CHANGES.md
@@ -17,6 +17,8 @@
 * Runtime: allow dynlink of precompiled js with separate compilation (#1676)
 * Lib: Modify Typed_array API for compatibility with WebAssembly
 * Compiler: improved global dead code elimination (#2206)
+* Compiler: add support for the Wasm backend in parts of the pipeline, in
+  prevision for the merge of wasm_of_ocaml
 
 
 ## Bug fixes